Move db.js to /utils
