+++ backend/auth/app.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const db = require('./db');
+const db = require('./utils/db');
 
 // Express configuration
 const app = express();
