+++ backend/auth/model/session.model.js
@@ -1,4 +1,4 @@
-const db = require('../db');
+const db = require('../utils/db');
 const { uuid } = require('uuidv4');
 const dotenv = require('dotenv/config');
